Refactory Open ID Functionality
